requirements.txt
numpy==1.22.3
Flask==2.0.2
loguru==0.5.3
python-dotenv==0.21.0
Flask-Cors==3.0.10
